Python/Flyordie/requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
Pillow==9.4.0
pyasn1==0.4.8
requests==2.28.2
rsa==4.9
soupsieve==2.3.2.post1
termcolor==2.2.0
urllib3==1.26.14
xmltodict==0.13.0
